requirements.txt
flask==2.0.1
pdfkit==0.6.1
flask-httpauth==4.4.0
python-dotenv==0.19.1
Werkzeug==2.0.1
gunicorn==20.1.0